+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `big` and `huge` props on `IconBadge` component.
+- Fix: Update svg `ArrowIcon` component.
+- Fix: Update `ArrowIcon`on `Pagination` component.
 - Breaking changes: Add `huge` size and replace `big` by `huge` size on `BulletList`.
 - Fix: Update margin on `ContributionCard` component.